runtime: move simple_buffer code in to cc
